feat(desktop): utility widgets ×6 — settings move into in-feed gear popovers [REL-38]
Per widget (clock / timer / weather / sysmon / uptime / github), the
mechanical PR 6 transform:

- widgets/<id>/Settings.tsx: the ConfigPanel's hooks + Section rows
  moved verbatim (every TickerPinSection kept), minus ConfigPanelLayout
  chrome; reset becomes a bottom ResetButton row.
- FeedTab: comfort mode floats a GearMenu top-right (absolute right-3
  top-3 over a relative min-h-full column) mounting Settings; the feed
  body is untouched (renamed <X>FeedBody). Feed-side add flows
  (add-timezone, CitySearch, Kuma/repo setup) untouched.
- ConfigPanel: thin scroll-chrome wrapper around the same Settings
  component — the Configure page keeps working until teardown.

Gate: tsc clean, vitest 490/490, all five Playwright suites green.
