+++ desktop/src/widgets/clock/FeedTab.tsx
@@ -1,5 +1,8 @@
 import { Clock } from "lucide-react";
 import { WorldClock } from "./WorldClock";
+import { GearMenu } from "../../components/widget-bar/GearMenu";
+import { useShell } from "../../shell-context";
+import ClockSettings from "./Settings";
 import type { FeedTabProps, WidgetManifest } from "../../types";
 
 export const clockWidget: WidgetManifest = {
@@ -22,7 +25,31 @@ export const clockWidget: WidgetManifest = {
   FeedTab: ClockFeedTab,
 };
 
-function ClockFeedTab({ mode }: FeedTabProps) {
+function ClockFeedTab(props: FeedTabProps) {
+  return (
+    <div className="relative flex min-h-full flex-col">
+      {/* Comfort mode floats the gear top-right — the widget's settings
+          surface once the Configure page dies. */}
+      {props.mode === "comfort" && (
+        <div className="absolute right-3 top-3 z-10">
+          <ClockGear />
+        </div>
+      )}
+      <ClockFeedBody {...props} />
+    </div>
+  );
+}
+
+function ClockGear() {
+  const { prefs, onPrefsChange } = useShell();
+  return (
+    <GearMenu ariaLabel="Clock settings" panelClassName="right-0 w-80">
+      <ClockSettings prefs={prefs} onPrefsChange={onPrefsChange} />
+    </GearMenu>
+  );
+}
+
+function ClockFeedBody({ mode }: FeedTabProps) {
   return (
     <div className="p-3">
       <WorldClock compact={mode === "compact"} />
